+++ pkg/database/ytdb.go
@@ -38,7 +38,7 @@ func YtGetStatus(Group, Member int64, Status, Region string) ([]LiveStream, erro
 			}
 			defer rows.Close()
 
-		} else if Status == "past" {
+		} else if Status == config.PastStatus {
 			rows, err = DB.Query(`SELECT Youtube.* FROM Vtuber.Youtube Inner join Vtuber.VtuberMember on VtuberMember.id=VtuberMember_id Inner join Vtuber.VtuberGroup on VtuberGroup.id = VtuberGroup_id Where (VtuberGroup.id=? or VtuberMember.id=?) AND Status=? Order by EndStream DESC Limit ?`, Group, Member, Status, limit)
 			if err != nil {
 				return nil, err